requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Document Processing
unstructured==0.11.6
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2

# LLM Integration
ollama==0.1.6
langchain==0.1.0
langchain-community==0.0.10

# Astrology Calculations
pyswisseph==2.10.3.2
pytz==2023.3
geopy==2.4.1
timezonefinder==6.2.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Web Framework (Streamlit UI)
streamlit==1.29.0
plotly==5.18.0
streamlit-extras==0.3.6
streamlit-folium==0.15.1

# Authentication & Security
bcrypt==4.1.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Payment & QR Code
qrcode[pil]==7.4.2

# Utilities
requests==2.31.0
tqdm==4.66.1
jsonlines==4.0.0

# Development
pytest==7.4.3
black==23.12.1
flake8==6.1.0
